Bug 458975: Fix ContextMenuFinder returns disposed menu items
The method findMenus() should not iterate through the shells of the
display when trying to find menu items. Doing so may dispose the
previous menu and its items at each iteration if the context menu is a
dynamic menu (e.g. MenuManager.setRemoveAllWhenShown(true)).

In the scope of a ContextMenuFinder there is only one menu associated
with the control, it should be queried directly.

Change-Id: Icfc5e05bf083626304e5a77054535352788b7f4b
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>