Support dynamic menus using JFace MenuManager in MenuTab example
Button "Menu Manager" is added to create menus using the JFace
MenuManager. Button "Dynamic" is added to make it dynamic by using
setRemoveAllWhenShown(true) and an IMenuListener.

This will help to test a fix for Bug 261360. The popup menu unit tests
fail if a dynamic menu manager is used.

The NO_RADIO_GROUP option is not supported with the JFace MenuManager.
Otherwise, the option handling is fixed to properly apply to sub-menus.

Change-Id: Ib56ff63351f47ba236b1f6ba37d4097fbc9dc949
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>