Test dynamic menus in SWTBotPopupMenuTest
Each unit test is now run twice, once with static menus and once with
dynamic menus.

The MenuTab example is updated to hook listeners on the shell. It is
updated to use a new Push action instance every time the dynamic menu is
populated (this allows testing that a new MenuItem is created and the
old MenuItem is disposed). It is also fixed so that sub-menus properly
support the Dynamic option, and to replace the use of recent method
MenuManager.setImageDescriptor() which is not available for all targets.

The targets are updated to use JUnit 4.12 for the parameterized test.

Change-Id: I36547154c8e7fad5fcbfbd3221f11f89c4f84261
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>