Speed up SWTBot tests in org.eclipse.swtbot.eclipse.finder.test package
Instead of waiting for the Welcome view (which can time out) and closing
it if it is found, iterate through all views and close the Welcome view
if it is part of the list.

Instead of waiting for the Window > Open Perspective > Java menu (which
can time out) and clicking it if it is found, check first if the active
perspective is not the Java perspective.

Make all test classes in org.eclipse.swtbot.eclipse.finder.test extend
AbstractSWTBotEclipseTest to reuse its @BeforeClass and @After code.

Change-Id: Ie3aec0d45c6e2ce5ca304ed8f45e21ac7549b226
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>