Use 'planner' include mode in target definitions
Since Luna, the Platform IDE feature no longer includes some ECF
plug-ins (e.g. org.eclipse.ecf and more), but it includes the plug-in
org.eclipse.equinox.p2.transport.ecf which depends on those ECF plug-ins
that are not part of our target definitions.

Instead of manually figuring out which plug-in or feature needs to be
included to resolve the dependency, we can use the include mode
'planner' instead of 'slicer', which will automatically resolve it.

This will prevent some validation errors in Run Configurations and
exceptions being thrown during SWTBot unit tests.

Change-Id: I4d26ef1134674b3ce2c199c2ab15b07cee8542a7