Bump swtbot-baseline.target sequence number
Because the contents of the target file have not changed but the
contents of the repositories it references have, the sequence number
needs to be increased for Eclipse to pick up the changes. This needs to
be done manually by the user by editing the API baseline in Window >
Preferences > Plug-in Development > API Baselines and pressing Refresh.

Change-Id: Id50ff403e14cfb65a906114eabf8cece42e4ff09