Bug 506628: Create SWTBot recorder server/client
Adds a server for recording SWTBot code and a client for viewing
generated code. The server runs in one Eclipse instance where tests
are recorded and the client runs in a separate Eclipse instance where
tests cases can be written and edited.

This change makes it possible to simultaneously have one workspace
open where a test is being recorded and another workspace where test
cases are being written.

Change-Id: I7b071498a143c210d4ed7692d28fe1e99562a0cf
Signed-off-by: Isaac Arvestad <isaac.arvestad@ericsson.com>