Add property to control GTK version
The tests seem to take forever when using GTK3.
See bug 506807.
With the new flag added by this patch, one can force GTK3/GTK2 when desired.

This change also cleans pulls up a common property for VM args.
It also adds an environment variable to workaround a crash in Ubuntu.

Change-Id: Ie740e13e70edccae4dff3a81a481acbc58a717d4
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>