Bug 337548: Fix SWTBotLink.click()
The method SWTBotLink.click() is changed to reliably click on the widget
and select the first hyperlink in the receiver's text, if there is any.

The method no longer posts mouse events to perform the click, and will
not interfere with the mouse cursor.

The method SWTBotLink.click(int) is added to click on the widget and
select the hyperlink at the specified index, if there is any.

Change-Id: I918aed40ae438fd8e5e1486ca162dd44786ec809
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>