Bug 546259: SWTBotEditor.isActive() returns true when not active part
SWTBotEditor.isActive() now returns true only if the editor is the
active part.

SWTBotEditor.isActiveEditor() is added and returns true if the editor is
the active editor, even if it is not the active part.

SWTWorkbenchBot.activePart() is added and returns the active part, which
may be a view or an editor.

Change-Id: I696da9f9462b8dbeca51ea5a0b4ca52142b81a66
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>