Bug 546259: SWTBotEditor.isActive() returns true when not active part
Revert the changes in https://git.eclipse.org/r/140326

The Javadoc of SWTBotEditor.isActive() is updated to reflect that it
returns true if the editor is the active editor, even if it is not the
active part.

SWTBotEditor.isActivePart() is added and returns true if the editor is
the active part.

Change-Id: Icaf3cf85698d1e1b81a29c4d0f5ab1937bfb79ea
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>